requirements.txt
PyQt5>=5.15.11
requests>=2.28.0
PyYAML>=6.0
pyinstaller>=5.13.0
pydantic~=2.10.5